[Refactor] Move network_util into base module
