+++ be/src/agent/heartbeat_server.cpp
@@ -44,6 +44,7 @@
 
 #include "agent/master_info.h"
 #include "agent/task_worker_pool.h"
+#include "base/network/network_util.h"
 #include "common/process_exit.h"
 #include "common/status.h"
 #include "common/system/cpu_info.h"
@@ -52,7 +53,6 @@
 #include "service/backend_options.h"
 #include "storage/storage_engine.h"
 #include "util/debug_util.h"
-#include "util/network_util.h"
 #include "util/thrift_server.h"
 
 using std::fstream;
