+++ be/src/exec/pipeline/exec_state_reporter.cpp
@@ -20,10 +20,10 @@
 #include <memory>
 
 #include "agent/master_info.h"
+#include "base/network/network_util.h"
 #include "runtime/client_cache.h"
 #include "runtime/exec_env.h"
 #include "service/backend_options.h"
-#include "util/network_util.h"
 #include "util/starrocks_metrics.h"
 #include "util/thrift_rpc_helper.h"
