+++ be/src/service/backend_options.cpp
@@ -20,12 +20,12 @@
 #include <algorithm>
 #include <ostream>
 
+#include "base/network/cidr.h"
+#include "base/network/network_util.h"
 #include "common/config.h"
 #include "common/logging.h"
 #include "common/status.h"
 #include "gutil/strings/split.h"
-#include "util/cidr.h"
-#include "util/network_util.h"
 
 namespace starrocks {
