+++ be/src/util/thrift_rpc_helper.cpp
@@ -36,13 +36,13 @@
 
 #include <sstream>
 
+#include "base/network/network_util.h"
 #include "base/time/monotime.h"
 #include "common/status.h"
 #include "gen_cpp/FrontendService.h"
 #include "runtime/client_cache.h"
 #include "runtime/exec_env.h"
 #include "runtime/runtime_state.h"
-#include "util/network_util.h"
 #include "util/thrift_util.h"
 
 namespace starrocks {
